module refactoring

+++ src/post/post.controller.ts
@@ -29,10 +29,7 @@ export class PostController {
   ): Promise<string> {
     /** 글 등록 */
     await this.postService.createPost(req.user.email, post);
-
-    /** 알림 생성 및
-     *  등록된 태그 보유 account 모두 알림 등록 */
-    return;
+    return 'post save successful';
   }
 
   /** Get Posts */
@@ -74,6 +71,7 @@ export class PostController {
   async deletePost( @Request() request, @Param('postId') postId: string): Promise<string> {
     // 본인의 게시글인지 확인
     const isCheck = await this.postService.checkIsMine(request.user.email);
+    if (!isCheck) throw new ForbiddenException();
     // 삭제 진행
     await this.postService.deletePost(postId);
     return 'post delete successful';
